Move problem data input from config file to directory
Previously, to specify problem data (e.g. system dynamics, initial state and input bounds, contraction matrix),
you would enter code snippets into the config file. While this is fine (and quite convenient) for systems with problem
data that can be easily described in a couple of lines of code, it makes it very difficult to describe complicated systems,
and isn't very flexible in general. Now, the user provides this problem data by filling out several .cl files in the same
directory as the config file. The advantage here is that the user can specify the problem data directly in code, rather
than out-of-context code snippets, and can use as many lines as they like, instead of just one, and in general they have
much more freedom than before. As the code is now, you can only specify problem data in the new way. It would be nice if
the user chould choose which of the two ways to input problem data, but I'll handle that later.